Changed percent for sapphire tools from int to float

+++ src/main/java/net/trique/mythicupgrades/item/base/VirtualSapphireTool.java
@@ -1,7 +1,7 @@
 package net.trique.mythicupgrades.item.base;
 
 public interface VirtualSapphireTool {
-    void setPercent(int percent);
+    void setPercent(float percent);
 
-    int getPercent();
+    float getPercent();
 }
